:recycle: [Root App] Move to /components
