+++ src/components/app/App.css
@@ -1,10 +1,12 @@
 @import url('https://rsms.me/inter/inter-ui.css');
 
 html {
+  font-family: 'Inter UI', sans-serif;
+
+  margin: 0;
+  padding: 0;
   height: 100%;
   width: 100%;
-
-  font-family: 'Inter UI', sans-serif;
 }
 
 @supports (font-variation-settings: normal) {
