+++ src/components/root/index.style.js
@@ -1,0 +1,10 @@
+import STYLE from 'constants/style'
+
+const style = {
+  base: {
+    ...STYLE.ABSOLUTE_FILL,
+    backgroundColor: STYLE.COLOR.WHITE,
+  },
+}
+
+export default style
